Make migrating fill in the gaps.
It often happens when developing that migrations are committed
out of order. This fixes goose so that it can do the right thing
in that scenario:

If asked to migrate 'up' then run all the up migrations that are older
than the target migration which have not been migrated up and then run
the target itself.

If asked to migrate 'down' then run all down migrations that are
newer than the target migration which have been migrated up, but don't
run the down migration of the target itself.

The old behaviour was to:

If the target is newer than current, run all the up migrations between
the two.

If the target is older than urrent, run all the down migrations between
the two.